+++ packages/ponder/ponder.schema.ts
@@ -35,7 +35,7 @@ export const job = onchainTable(
 		acceptedAt: t.bigint(),
 		finishedAt: t.bigint(), // When the job was finished
 		canceledAt: t.bigint(), // When the job was canceled
-		rating: t.bigint(),
+		rating: t.integer(), // Rating given by the client to the freelancer
 		clientReceived: t.boolean().notNull(),
 		freelancerDelivered: t.boolean().notNull(),
 		lastTransactionHash: t.varchar({ length: 256 }).notNull(),
